backend/requirements.txt


Flask==3.1.2
flask-cors==6.0.1
Flask-PyMongo==3.0.1
flatbuffers==25.2.10
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
pillow==11.2.1
pymongo==4.14.1
python-dotenv==1.1.1
